fix: change syscall type to u32 #3

+++ syslog-common/src/lib.rs
@@ -3,9 +3,9 @@
 #[repr(C)]
 #[derive(Clone, Copy)]
 pub struct SysCallLog {
-    pub ts      : u64,
-    pub syscall : u64,
-    pub pid     : u32,
+    pub ts            : u64,
+    pub syscall       : u32,
+    pub pid           : u32,
     pub pname_bytes   : [u8; 16]
 }
